requirements.txt
# requirements.txt: Dependências do projeto
streamlit==1.29.0
simpy==4.1.1
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
pandas==2.2.2
reportlab==4.2.2
chartjs>=0.2.1
plotly-6.1.0

